Allow PlaceObject3 tag as child of DefineSprite tag.
from https://bugs.launchpad.net/swfmill/+bug/683793:

When importing a SWF in simple mode, swfmill places most of the objects into a DefineSprite tag, then places that on the main timeline with the requested name. However, it does not place any objects to which filters have been applied into this DefineSprite tag. Instead, these objects are placed in the root timeline of the resulting movieclip. This has the net result of making these objects "ignore" the playhead position of the imported movie clip, and also their instance names are in a different scope than their un-filtered bethren.

Closer inspection shows that objects with filters are placed with a relatively new PlaceObject3 tag; the same object without any filters is placed with a PlaceObject2 tag. However Adobe's own SWF format documentation does not list PlaceObject3 as an allowable control tag inside a DefineSprite tag. Therefore, this "shortcoming" in swfmill is not a true bug, because it's following spec: putting PlaceObject2 tags inside a DefineSprite tag, but leaving PlaceObject3 tags outside of it. However, the Flash IDE (at least v9/CS3) does put PlaceObject3 tags inside DefineSprite tags, despite the spec, and the resulting SWF plays properly. Because of this, I suspect that the SWF format documentation simply needs to be updated to include PlaceObject3 as an allowable tag. The current list of allowable tags may well have predated the advent of PlaceObject3.

Anyway, because of the structure of swfmill, the fix for this issue is trivial, and I suppose the only question might be whether or not to do this at all, since it deviates from the vendor's spec.

I've attached a patch for consideration. I have already applied it to my checkout, built it successfully, and found that it solves this problem without creating new ones (yet, anyway).