#22: Fix incorrect TTF font metrics.
On x64 (and possibly other platforms), the metrics were being set to arbitrary values picked out of the wrong bytes in a struct.

It's not clear whether we should really be using usWinAscent and usWinDescent in preference to sTypoAscender and sTypoDescender, but that's what we were doing before on the platforms that were "working" so we'll preserve that behaviour for now.